@@ -1,24 +1,24 @@
 class GroceryData{
-   static List<Map<String, dynamic>> groceryList = [
-      {
-        'name': 'Apples',
-        'quantity': 5,
-        'price': 2.99,
-      },
-      {
-        'name': 'Bananas',
-        'quantity': 8,
-        'price': 1.99,
-      },
-      {
-        'name': 'Milk',
-        'quantity': 2,
-        'price': 3.49,
-      },
-      {
-        'name': 'Bread',
-        'quantity': 1,
-        'price': 2.29,
-      },
-    ];
+  static List<Map<String, dynamic>> groceryList = [
+    {
+      'name': 'Apples',
+      'quantity': 5,
+      'price': 2.99,
+    },
+    {
+      'name': 'Bananas',
+      'quantity': 8,
+      'price': 1.99,
+    },
+    {
+      'name': 'Milk',
+      'quantity': 2,
+      'price': 3.49,
+    },
+    {
+      'name': 'Bread',
+      'quantity': 1,
+      'price': 2.29,
+    },
+  ];
 }
